Extract tests into dedicated crate
This allows us to manage the harness code as a separate crate and
also further simplifies tearing apart the `daemon` crate into `maker`
and `taker`.
