+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["daemon", "rocket-basicauth", "rust-embed-rocket", "taker", "maker", "shared-bin"]
+members = ["daemon", "rocket-basicauth", "rust-embed-rocket", "taker", "maker", "shared-bin", "daemon-tests"]
 resolver = "2"
 
 [patch.crates-io]
