+++ daemon-tests/src/mocks/monitor.rs
