+++ daemon-tests/src/mocks/wallet.rs
@@ -1,18 +1,19 @@
 use anyhow::Result;
-use bdk::bitcoin::ecdsa;
-use bdk::bitcoin::util::psbt::PartiallySignedTransaction;
-use bdk::bitcoin::Amount;
-use bdk::bitcoin::Txid;
-use bdk::wallet::tx_builder::TxOrdering;
-use bdk::wallet::AddressIndex;
-use bdk::FeeRate;
+use daemon::bdk;
+use daemon::bdk::bitcoin::ecdsa;
+use daemon::bdk::bitcoin::util::psbt::PartiallySignedTransaction;
+use daemon::bdk::bitcoin::Amount;
+use daemon::bdk::bitcoin::Txid;
+use daemon::bdk::wallet::tx_builder::TxOrdering;
+use daemon::bdk::wallet::AddressIndex;
+use daemon::bdk::FeeRate;
 use daemon::bdk_ext::new_test_wallet;
+use daemon::maia::secp256k1_zkp::Secp256k1;
+use daemon::maia::PartyParams;
+use daemon::maia::TxBuilderExt;
 use daemon::model::Timestamp;
 use daemon::model::WalletInfo;
 use daemon::wallet;
-use maia::secp256k1_zkp::Secp256k1;
-use maia::PartyParams;
-use maia::TxBuilderExt;
 use mockall::*;
 use rand::thread_rng;
 use std::sync::Arc;
